Refactor structure and fix test
Convolution files are now in own subdiretory.
Convolution main is refacotred in two files: 2d-convolution.c and
convolution-run.c.
This was neccessary in order to test the real kernel application,
run_on_padded_image and so on.

Also new subdirectory for nbody, it is also split into util and main
files.

Integration test of the function run_on_padded_image.